# levenshtein zig: handle arbitrary long input strings