@PEZ 2025-01-10
Thanks! 🙏 

Seems your changes also fixed something with an unclean exit:

**Before**:

```
Checking Zig
Check passed
Benchmarking Zig
Benchmark 1: ./zig/code aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa ...
ciladucljdamhnafrxyabwuihrusmthypjxormffegjwxioyztsdwqemzjdf ...
  Time (mean ± σ):     595.0 ms ±   6.1 ms    [User: 574.7 ms, System: 4.8 ms]
  Range (min … max):   586.9 ms … 603.4 ms    7 runs
 
  Warning: Ignoring non-zero exit code.
```

**After**:

```
Checking Zig
Check passed
Benchmarking Zig
Benchmark 1: ./zig/code aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa ...
ciladucljdamhnafrxyabwuihrusmthypjxormffegjwxioyztsdwqemzjdf ...
  Time (mean ± σ):     589.3 ms ±   2.5 ms    [User: 578.3 ms, System: 4.9 ms]
  Range (min … max):   585.7 ms … 592.8 ms    7 runs
```

Note to self (and @bddicken): We should be able to pick up non-zero exit codes in the correctness check. Running benchmarks for programs not exiting cleanly doesn't make sense.